feat(admin): add configurable maximum file upload size
