+++ packages/admin/admin-bundler/src/types.ts
@@ -1,7 +1,7 @@
 import { AdminOptions } from "@medusajs/types"
 
 export type BundlerOptions = Required<Pick<AdminOptions, "path">> &
-  Pick<AdminOptions, "vite" | "backendUrl" | "storefrontUrl"> & {
+  Pick<AdminOptions, "vite" | "backendUrl" | "storefrontUrl" | "maxUploadFileSize"> & {
     outDir: string
     sources?: string[]
     plugins?: string[]
