+++ packages/admin/admin-bundler/src/utils/config.ts
@@ -54,6 +54,7 @@ export async function getViteConfig(
       __AUTH_TYPE__: JSON.stringify(authType),
       __JWT_TOKEN_STORAGE_KEY__: JSON.stringify(jwtTokenStorageKey),
       __STOREFRONT_URL__: JSON.stringify(storefrontUrl),
+      __MAX_UPLOAD_FILE_SIZE__: JSON.stringify(options.maxUploadFileSize ?? 1024 * 1024),
     },
     server: {
       fs: {
