+++ packages/admin/dashboard/src/lib/format-file-size.ts
@@ -1,0 +1,19 @@
+export function formatFileSize(
+  bytes: number,
+  decimalPlaces: number = 2
+): string {
+  if (bytes === Infinity) {
+    return "unlimited"
+  }
+  if (bytes === 0) {
+    return "0 Bytes"
+  }
+
+  const k = 1024
+  const sizes = ["Bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"]
+  const i = Math.floor(Math.log(bytes) / Math.log(k))
+
+  return (
+    parseFloat((bytes / Math.pow(k, i)).toFixed(decimalPlaces)) + " " + sizes[i]
+  )
+}
