+++ packages/admin/dashboard/src/vite-env.d.ts
@@ -18,3 +18,4 @@ declare const __STOREFRONT_URL__: string | undefined
 declare const __BASE__: string
 declare const __AUTH_TYPE__: "session" | "jwt" | undefined
 declare const __JWT_TOKEN_STORAGE_KEY__: string | undefined
+declare const __MAX_UPLOAD_FILE_SIZE__: number | undefined
